+++ CHANGELOG.md
@@ -16,6 +16,42 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - N/A
 
+## [1.10.1] 2026-06-06
+This version add the test for the next enspoints:
+- orders:
+  - /create/
+  - /update/{order_id}
+  - /delete/{order_id}
+  - /get/{order_id} # created in this version.
+  - /get/all/ # created in this version
+- items:
+  - GET /all/
+  - POST /{order_id}/items
+  - /delete/{order_id}/items/{item_id}
+- products:
+  - /all/
+- flavors:
+  - /all/
+
+### Added
+- enpoint for get an order by id in `src/api/v1/apps/orders/router.py`
+- enpoint for get all orders in `src/api/v1/apps/orders/router.py`
+- Tests:
+  - get and post orders endpoints in `tests/integration/orders/test_order_post_get_endpoints.py`
+  - patch and delete endpoints in `test/integration/orders/test_orders_update_and_delete_endpoints.py`
+  - get and post items endpoints in `tests/integration/orders/items/test_items_post_get_endpoints.py`
+  - patch and delete items endpoints in `test/integration/orders/items/test_items_update_and_delete_endpoints.py`
+
+### Changed
+- Schemas to create the correct flow for order details in `src/api/v1/apps/orders/schemas.py`
+- Added an all endpoint test for products in `test/integration/products/test_products.py`
+- Added an all endpoint test for flavors in `tests/integration/flavors/test_flavor_post_get_endpoints.py`
+
+### Fixed
+- Added the user_id to the current_user payload for fixture auth user in `tests/conftest.py`
+- Fixed the missing "/" in tests in `tests/integration/test_update_and_delete_products_endpoints.py`
+- Removed the += in nested quantity update in `src/api/v1/apps/orders/services.py`
+
 ## [1.10.0] 2026-06-05
 This version add the orders module